+++ spec/acceptance/api/v2/constants_spec.rb
@@ -93,7 +93,7 @@ resource 'Constants' do
 
         expect(resource_object['alert_frequencies']).to be_a_hash_with_types String, String
 
-        expect(resource_object['csv_import'].keys.size).to eq 5
+        expect(resource_object['csv_import'].keys.size).to eq 4
 
         expect(resource_object['tnt_import'].keys.size).to eq 1