+++ spec/controllers/api/v2/account_lists/imports/csv_controller_spec.rb
@@ -10,8 +10,15 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
   let!(:id) { import.uuid }
   let!(:parent_param) { { account_list_id: account_list_id } }
   let!(:imports) do
-    create_list(:csv_import, 2, account_list: account_list, user: user, in_preview: true,
-                                file_headers_mappings: file_headers_mappings, file_constants_mappings: file_constants_mappings)
+    create_list(
+      :csv_import,
+      2,
+      account_list: account_list,
+      user: user,
+      in_preview: true,
+      file_headers_mappings: file_headers_mappings,
+      file_constants_mappings: file_constants_mappings
+    )
   end
   let!(:import) { imports.first }
   let!(:resource) { import }
@@ -19,8 +26,8 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
   let(:file_headers_mappings) do
     {
       'city' => 'city',
-      'commitment_amount' => 'amount',
-      'commitment_frequency' => 'frequency',
+      'pledge_amount' => 'amount',
+      'pledge_frequency' => 'frequency',
       'contact_name' => 'fname',
       'country' => 'country',
       'email_1' => 'email-address',
@@ -44,15 +51,15 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
 
   let(:file_constants_mappings) do
     {
-      'status' => {
-        'partner_financial' => 'Praying and giving'
-      },
-      'commitment_frequency' => {
-        '1_0' => 'Monthly'
-      },
-      'newsletter' => {
-        'both' => 'Both'
-      }
+      'status' => [
+        { 'id' => 'Partner - Financial', 'values' => ['Praying and giving'] }
+      ],
+      'pledge_frequency' => [
+        { 'id' => 'Monthly', 'values' => ['Monthly'] }
+      ],
+      'newsletter' => [
+        { 'id' => 'Both', 'values' => ['Both'] }
+      ]
     }
   end
 
@@ -145,14 +152,18 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
 
     it 'defaults source to csv' do
       api_login(user)
-      post :create, full_correct_attributes.tap { |params| params[:data][:attributes][:source] = 'twitter' }
+      post :create, full_correct_attributes.tap { |params|
+        params[:data][:attributes][:source] = 'twitter'
+      }
       import = Import.find_by_uuid(JSON.parse(response.body)['data']['id'])
       expect(import.source).to eq 'csv'
     end
 
     it 'defaults in_preview to true' do
       api_login(user)
-      post :create, full_correct_attributes.tap { |params| params[:data][:attributes].delete([:in_preview, 'in_preview']) }
+      post :create, full_correct_attributes.tap { |params|
+        params[:data][:attributes].delete([:in_preview, 'in_preview'])
+      }
       import = Import.find_by_uuid(JSON.parse(response.body)['data']['id'])
       expect(import.in_preview?).to eq true
     end
@@ -167,7 +178,9 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
 
     it 'persists cached file data' do
       api_login(user)
-      post :create, full_correct_attributes.tap { |params| params[:data][:attributes].delete([:in_preview, 'in_preview']) }
+      post :create, full_correct_attributes.tap { |params|
+        params[:data][:attributes].delete([:in_preview, 'in_preview'])
+      }
       import = Import.find_by_uuid(JSON.parse(response.body)['data']['id'])
       expect(import.file_headers).to be_present
       expect(import.file_headers).to be_a Hash
@@ -175,17 +188,6 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
       expect(import.file_constants).to be_a Hash
       expect(import.file_row_samples).to be_present
       expect(import.file_row_samples).to be_a Array
-    end
-
-    it 'includes sample contacts' do
-      api_login(user)
-      post :create, (full_correct_attributes.tap do |params|
-        params[:data][:attributes].delete([:in_preview, 'in_preview'])
-        params[:include] = 'sample_contacts'
-      end)
-      included = JSON.parse(response.body)['included']
-      expect(included.size).to eq 1
-      expect(included.first['type']).to eq 'contacts'
     end
   end
 
@@ -201,9 +203,12 @@ describe Api::V2::AccountLists::Imports::CsvController, type: :controller do
     it 'permits all params under file_constants_mappings' do
       api_login(user)
       put :update, full_correct_attributes.tap { |params|
-        params[:data][:attributes][:file_constants_mappings] = { 'testing' => { 'nested' => '1234' } }
+        params[:data][:attributes][:file_constants_mappings] =
+          { 'testing' => [{ 'id' => 'nested', 'values' => ['1234'] }] }
       }
-      expect(import.reload.file_constants_mappings).to eq('testing' => { 'nested' => '1234' })
+      expect(import.reload.file_constants_mappings).to eq(
+        'testing' => [{ 'id' => 'nested', 'values' => ['1234'] }]
+      )
     end
 
     it 'permits assigning empty hashes to the mappings' do