+++ spec/services/constant_list_spec.rb
@@ -69,9 +69,6 @@ RSpec.describe ConstantList, type: :model do
     it { expect(subject.csv_import[:constants].keys & subject.csv_import[:supported_headers].keys).to eq subject.csv_import[:constants].keys }
     it { expect(subject.csv_import[:constants].keys).to eq CsvImport.constants.keys }
     it { expect(subject.csv_import[:max_file_size_in_bytes]).to eq Import::MAX_FILE_SIZE_IN_BYTES }
-    it { expect(subject.csv_import[:constants_from_top_level]).to be_an Hash }
-    it { expect(subject.csv_import[:constants_from_top_level].keys.collect(&:to_s)).to eq subject.csv_import[:constants].keys }
-    it { expect(subject.csv_import[:constants_from_top_level].keys.collect(&:to_s)).to eq CsvImport.constants.keys }
   end
 
   describe '#tnt_import' do