+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 17 November 2017
+- CSV Import no longer accepts hash key value format
+
 ## 8 November 2017
 - designation account display name field no longer duplicates designation_number
 - donor account display name field no longer duplicates account_number